feat(error-template-support): add support for response error templates
This commit bears the support for response error templates which are resolved at runtime in the response handler class. Other than the feature, this commit contains unit tests, a fix for the default case at the endpoint level along with some refactoring of the existing code in error case and response handler classes.

closes #24